rename sleep.sh => start-up.sh
sleep.sh started out just being a script to call to `sleep`, which
kept our containers running so we could hopon to them. However
over the course of our explorations, this file started being built
out to something much more than just `sleep`. Changing the name
reflects that this file is all of the settings needed to start-up
our containers with the expected configurations that make our
little internetwork run.

Co-authored-by: Ashish <adixit@fastly.com>
Co-authored-by: Peba <peba@fastly.com>