remove creating a new bridge network
We did a bit of a run-around when we initially wrote this chapter. I believe we thought that removing the default network and adding a new bridge network would give us more control... But... that just made this chapter more about how to navigate docker than about how to create a simple network... This commit removes that process and just manages the default network docker creates which gets us to the same point with more information and less run-around.