update main readme and add chapter 000
there's a lot to learn in these chapters. we don't want docker to
be an inadventant gatekeeper to folks who are interested in
networking but don't have experience with docker yet. adding a
simple prequel to explain the setup and what its doing helps to
minimize the impact of that barrier to entry.