cleanup: Updates hopon and restart
As we embark on explorations being outside of chapters folder, we needed a way
to deterministically name images, containers and networks so they could be
brought down and back up with our restart scripts and we could hopon to them as
needed. This change only constitutes the changes to the scripts with future
commits updating docker compose files for chapters adding a more deterministic
naming of various objects.

Co-authored-by: Peba <peba@fastly.com>
Co-authored-by: Squee Mueller <squeemishly@gmail.com>