switch to using avahi for cross network name resolution
it turns out, the entire internet is aghast at the idea of performing name
resolution across networks using multicast DNS. However, the approved way
of doing it to use some gateway. Avahi has that gateway built in.

In order to take advantage of the gateway, we needed to enable some reflector
settings:
- enable-reflector=yes
- reflect-ipv=yes

Once those were turned on, mDNS was easily perfomed. We captured the tcpdump
output from 3 machines on 2 networks to understand the process of name
resolution that was being performed.

Next steps: complete the documentation of this chapter, including why we
went with avahi over a more bespoke solution.

Co-authored-by: Ashish <adixit@fastly.com>
Co-authored-by: Peba <peba@fastly.com>