docker(network): define rp_filter at network level
Ashish was running into issues getting the byoi containers running because of
an error that we did not see with others. We isolated the issue to the
recommendation from the error message and moving the definitions for rp_filter
to the network definitions seems to have solved it.  # Please enter the commit
message for your changes. Lines starting

There is some question on whether or not this is truly applying the rp_filter.