ergonomics: Containers get vimrc
During the run through of the mild lab, we noticed that editing things using
vim within the various container could be improved with the base vimrc. This
adds that to all the different types of containers in our little internet so we
have a decent vimrc default to work with.