update chapter 1.2 with new network maps and less docker work
This change started with replacing the old ascii network map with the new image
network map. In the process of doing that, I needed to update all of the
machine names. At that point, it meant we were starting with an internet that
only had a client and a router. That don't make no sense.

I decided to make our internet start with all networks and machines fully added
and ready to go. That wasn't the interesting part of the chapter anyway. This
change also means that our readers won't need to do anything with docker or
docker compose in this chapter, and I think we can all agree that that's a win.