embed all image links into markdown link
In order to allow users viewing the documentation to easily click on a
an image to get the full-size variant, embed all image links into a
markdown link.

Include title (alternate text) for each of the images, which can (and
will) be used as hover text in (most) presentational UIs.

Tested with GFM Markdown.