improve image handling: link reference definitions
Use link reference definitions instead of incorrectly nesting typical Markdown links into image definitions.

Link reference definitions are akin to footnotes. Each link reference
definition has a name, and can be placed anywhere in a file though they
usually are placed at the end.

In the flow of text, a link can be used (inserted) using [brackets].
The Markdown processor inserts the link (and optional title text) from
the definition.  There are several useful forms of reference links, both
inline in text and in definitions.

These links can be used in the same way that typical links and autolinks
are used, and they allow the visual separation of the use of the link
from the definition of the link.

See: https://spec.commonmark.org/0.31.2/#link-reference-definitions

Within complex document sets, reference links provide convenient
management opportunity for interlinking.  Link destinations (i.e. to
other files) can all be kept in the same section of the file, slightly
easing the maintenance burden when moving files around. Because Markdown
parsers do not render reference link definitions into the output, they
can reside comfortably at the bottom of the file.

This is a better way of providing the image-handling technique that
worked despite being invalid Markdown (see 0cd7e08 and 664bfe6d). The
`pymarkdown` linter approves of this new approach (esspied in the
documentation of the mkdocs project main README file).

Bonus:  In addition to switching all of the files to use this technique
for the presentation of the images, see also a demonstrate the use of
reference links a bit more completely in chapter 1.1, showing how to use
them for external links.

The verb "pronk" is pretty nifty.